refactor: move receipt body codec out of mysql
